frame: u32 -> usize

+++ src/attributes/attribute.rs
@@ -1,3 +1,3 @@
 pub trait Attribute<T> {
-    fn get_frame(&self, frame: u32) -> T;
+    fn get_frame(&self, frame: usize) -> T;
 }
